+++ slog/cmd/main.go
@@ -1,10 +1,11 @@
 package main
 
 import (
-	"encoding/json"
 	"log"
 	"log/slog"
 	"net/http"
+
+	"github.com/tamaco489/go_sandbox/slog/internal/handler"
 )
 
 func main() {
@@ -12,7 +13,7 @@ func main() {
 	port := ":8080"
 
 	// Routing setup
-	http.HandleFunc("/api/v1/health", handleHealth)
+	http.HandleFunc("/api/v1/health", handler.HandleHealth)
 
 	// Start server
 	slog.Info("Server started", "port", port)
@@ -20,22 +21,4 @@ func main() {
 	if err := http.ListenAndServe(port, nil); err != nil {
 		log.Fatal("Server startup error:", err)
 	}
-}
-
-// HealthResponse represents the health check response structure
-type HealthResponse struct {
-	Message string `json:"message"`
-}
-
-// handleHealth handles health check API
-func handleHealth(w http.ResponseWriter, r *http.Request) {
-	w.Header().Set("Content-Type", "application/json; charset=utf-8")
-	w.WriteHeader(http.StatusOK)
-	response := HealthResponse{Message: "ok"}
-	json, err := json.Marshal(response)
-	if err != nil {
-		http.Error(w, err.Error(), http.StatusInternalServerError)
-		return
-	}
-	w.Write(json)
 }
